Moved StringSet to util package
